Make perpage behaviour more consistent
This adjusts the pagination code to allow returning fewer than the default number of items per page, and to reduce requests for more than the maximum to the maximum rather than the default.

The first bit of behaviour is desirable for me as in part of what I'm doing (trying to pick a random item) it's useful for me to be able to query the number of items matched by a query without initially returning any of the content of that query.
My current hack to achieve this is to set the page index to something out of range, (I'm usually using page 1 billion or so) which means I hit your database for a count and _only_ a count so it's blazing fast!

The latter just feels a bit more consistent with what an end-user would want! 😜

Additionally, this updates the documentation to be a bit more clear about the behaviour - as it was I was under the impression only 40 and 100 were valid page size values, but it turns out it was any page size between 40 and 100! :P

Unfortunately I don't have access to Visual Studio to verify this change, so I'd advise giving it a good double-check with some zero-item queries before deploying! 💜